Add files via upload
Consumer Analysis is a detailed analysis of a company’s ideal customers. It helps a business to better understand its customers and makes it easier for them to modify products according to the specific needs, behaviors and concerns of different types of customers. In this, I am going to use Python to analyze the customer purchase behavior in a grocery store.